requirements.txt
Django==5.0.7
django-cors-headers==4.3.0
tensorflow==2.15.0
numpy==1.26.0
requests==2.31.0
beautifulsoup4==4.12.2
google-api-python-client==2.95.0
python-dotenv==1.0.0